Add files via upload
Question 1 : Prompt user to input a sentence. The program will count all lower case alphabets, upper case alphabets, digits, and special
symbols. Then the program will display the all the counts on screen.

Question 2: The program accept four lists of four values. Then the program find the list whose sum of elements is the highest and display the list.

3. The program accept a number. If you reverse the given number and it is the same as the original number then the program should return true.

4. A program that find the L.C.M. of two input number.

5. A program that find the H.C.F of two input number.

6. The program that will accept two lists of integer. The program create a third list such that it contain only odd numbers from the first list and even numbers from the second list.

7. A program for the stock of a pharmacy, with menu:
         - 1. Add an item to stock pharmacy
         - 2. Remove item from stock
         - 3. Insert item at specified position.
The program perform all the operations. It uses list.

8. A program to make a two-player Rock-Paper-Scissors game.
It ask the player to input its choice, compare them, print out a message of congratulations to the winner, and ask if the players want to start a new game.